Lessons 159, 160, 161 and 162 - Texting and firing Employee
Lesson 159 - Texting Employees

It adds the `react-native-communications` module to send tex
to employee.

Lesson 160 - Modals as a reusable components

Lesson 161 - The Modal component continued

It creates the `Confirm` component to allow the app to reuse
it all over the place.

Lesson 162 - Styling the Modal

It adds some styling to the `Confirm` component.